neet-learning-assistant-phase1/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
openai
numpy
pdfminer.six
python-multipart
python-dotenv
